feat(signal-handling): Backports handling of SIGTERM to 1.2.0
SIGTERM is used in ECS to signal to tasks that they should exit
before being killed. Newer versions of `logbt` does this and more.

It is however not trivial to migrate ECS tasks
on a shared cluster to a newer `logbt` due to
https://github.com/mapbox/logbt/issues/37.

This code thus adds SIGTERM handling to the 1.2.0 version
for the situations where upgrading `logbt` is not feasible.